partly written Poisson alias table construction

+++ src/cu/poisson.cuh
@@ -7,11 +7,11 @@ namespace gplda {
 
 class Poisson {
   public:
-    size_t max_lambda;
-    size_t max_value;
+    int max_lambda;
+    int max_value;
     float** prob;
     float** alias;
-    Poisson(size_t ml, size_t mv);
+    Poisson(int ml, int mv);
     ~Poisson();
 };
